logger/requirements.txt
certifi
cffi
charset-normalizer
cryptography
idna
numpy
opencv-python
Pillow
pycparser
pynput
pywin32
requests
scipy
six
sounddevice
urllib3
keyboard
atexit
ctypes